requirements.txt
streamlit
PyMuPDF
pytesseract
Pillow
edge-tts
python-docx
pydub